♻️ move graphql context to UserService
