+++ users/src/main/kotlin/org/seasonist/resources/UserResource.kt
@@ -1,7 +1,6 @@
 package org.seasonist.resources
 
 import io.smallrye.graphql.api.Context
-import jakarta.json.JsonValue
 import jakarta.transaction.Transactional
 import org.eclipse.microprofile.graphql.*
 import org.eclipse.microprofile.graphql.GraphQLException.ExceptionType
@@ -16,28 +15,15 @@ class UserResource(
 	private val userService: UserService,
 	private val context: Context,
 ) {
-	companion object {
-		const val RECOMMENDATIONS_FIELD = "recommendations"
-		const val EXPERIENCES_FIELD = "experiences"
-	}
-
 	private val dateFormatter: SimpleDateFormat = SimpleDateFormat("yyyy-MM-dd")
 
 	@Query
 	@Description("Get User from an email")
-	fun getUserByEmail(@Name("email") email: String): User = this.userService.getUser(
-		email,
-		doesSelectedFieldsContains(RECOMMENDATIONS_FIELD),
-		doesSelectedFieldsContains(EXPERIENCES_FIELD)
-	)
+	fun getUserByEmail(@Name("email") email: String): User = this.userService.getUser(email, context)
 
 	@Query
 	@Description("Get User from an ID")
-	fun getUserById(@Name("userId") userId: UUID): User = this.userService.getUser(
-		userId,
-		doesSelectedFieldsContains(RECOMMENDATIONS_FIELD),
-		doesSelectedFieldsContains(EXPERIENCES_FIELD)
-	)
+	fun getUserById(@Name("userId") userId: UUID): User = this.userService.getUser(userId, context)
 
 	@Mutation
 	fun updateUserInformation(
@@ -53,11 +39,7 @@ class UserResource(
 	): User {
 		this.userService.updateUser(userId, firstname, lastname, email, phone, address, gender, bio, nationality)
 
-		val hasRecommendationsField = doesSelectedFieldsContains(RECOMMENDATIONS_FIELD)
-		val hasExperiencesField = doesSelectedFieldsContains(EXPERIENCES_FIELD)
-		val user = this.userService.getUser(userId, hasRecommendationsField, hasExperiencesField)
-
-		return user
+		return this.userService.getUser(userId, context)
 	}
 
 	@Mutation
@@ -70,9 +52,7 @@ class UserResource(
 		refereePhone: String?,
 		companyId: UUID?,
 	): User {
-		val hasRecommendationsField = doesSelectedFieldsContains(RECOMMENDATIONS_FIELD)
-		val hasExperiencesField = doesSelectedFieldsContains(EXPERIENCES_FIELD)
-		val user = this.userService.getUser(userId, hasRecommendationsField, hasExperiencesField)
+		val user = this.userService.getUser(userId, context)
 
 		val recommendation = Recommendation().apply {
 			this.firstname = refereeFirstname
@@ -83,7 +63,7 @@ class UserResource(
 			this.companyId = companyId
 		}
 		Recommendation.persist(recommendation)
-		if (hasRecommendationsField)
+		if (UserService.doesSelectedFieldsContains(UserService.RECOMMENDATIONS_FIELD, context))
 			user.recommendations.add(recommendation)
 
 		return user
@@ -95,15 +75,13 @@ class UserResource(
 		userId: UUID,
 		recommendationId: UUID,
 	): User {
-		val hasRecommendationsField = doesSelectedFieldsContains(RECOMMENDATIONS_FIELD)
-		val hasExperiencesField = doesSelectedFieldsContains(EXPERIENCES_FIELD)
-		val user = this.userService.getUser(userId, hasRecommendationsField, hasExperiencesField)
+		val user = this.userService.getUser(userId, context)
 
 		val recommendation = user.recommendations.find { it.id == recommendationId }
 			?: throw GraphQLException("Recommendation not found", ExceptionType.DataFetchingException)
 
 		Recommendation.delete("id", recommendationId)
-		if (hasRecommendationsField)
+		if (UserService.doesSelectedFieldsContains(UserService.RECOMMENDATIONS_FIELD, context))
 			user.recommendations.remove(recommendation)
 
 		return user
@@ -119,9 +97,7 @@ class UserResource(
 		description: String?,
 		companyId: UUID?,
 	): User {
-		val hasRecommendationsField = doesSelectedFieldsContains(RECOMMENDATIONS_FIELD)
-		val hasExperiencesField = doesSelectedFieldsContains(EXPERIENCES_FIELD)
-		val user = this.userService.getUser(userId, hasRecommendationsField, hasExperiencesField)
+		val user = this.userService.getUser(userId, context)
 
 		try {
 			val experience = Experience().apply {
@@ -133,7 +109,7 @@ class UserResource(
 				this.companyId = companyId
 			}
 			Experience.persist(experience)
-			if (hasExperiencesField)
+			if (UserService.doesSelectedFieldsContains(UserService.EXPERIENCES_FIELD, context))
 				user.experiences.add(experience)
 		} catch (e: ParseException) {
 			throw GraphQLException("Invalid date format", ExceptionType.ExecutionAborted)
@@ -145,29 +121,15 @@ class UserResource(
 	@Mutation
 	@Transactional
 	fun deleteExperience(userId: UUID, experienceId: UUID): User {
-		val hasRecommendationsField = doesSelectedFieldsContains(RECOMMENDATIONS_FIELD)
-		val hasExperiencesField = doesSelectedFieldsContains(EXPERIENCES_FIELD)
-		val user = this.userService.getUser(userId, hasRecommendationsField, hasExperiencesField)
+		val user = this.userService.getUser(userId, context)
 
 		val experience = user.experiences.find { it.id == experienceId }
 			?: throw GraphQLException("Experience not found", ExceptionType.DataFetchingException)
 
 		Experience.delete("id", experienceId)
-		if (hasExperiencesField)
+		if (UserService.doesSelectedFieldsContains(UserService.EXPERIENCES_FIELD, context))
 			user.experiences.remove(experience)
 
 		return user
 	}
-
-	private fun doesSelectedFieldsContains(fieldName: String): Boolean =
-		context.selectedFields.asJsonArray().any { field ->
-			if (!field.valueType.equals(JsonValue.ValueType.OBJECT))
-				return@any false
-
-			val obj = field.asJsonObject()
-			if (!obj.containsKey(fieldName))
-				return@any false
-
-			true
-		}
 }
